topicctl new action: rebalance
